Add files via upload
XAMPP server is used for local hosting. A database "login_sample_db" is used to store the signUp details.
loginstyle.css, signupstyle.css and background.png are present inside a folder "css"
For the beginning, Signup.php page needs to be used and the user needs to enter the username and password and then the user can login and go to the Index page where the user details, the chart and food ordering menu are available.